Add trace demo showing modifying the PC
* Use size_t for registers
* naked function so that we can get accurate PC
  (not going to work for AArch64)

+++ include/common/trace.h
@@ -10,31 +10,31 @@
 #error
 #else
 typedef struct {
-  uint32_t r0;
-  uint32_t r1;
-  uint32_t r2;
-  uint32_t r3;
-  uint32_t r4;
-  uint32_t r5;
-  uint32_t r6;
-  uint32_t r7;
-  uint32_t r8;
-  uint32_t r9;
-  uint32_t r10;
-  uint32_t r11;
-  uint32_t r12;
+  size_t r0;
+  size_t r1;
+  size_t r2;
+  size_t r3;
+  size_t r4;
+  size_t r5;
+  size_t r6;
+  size_t r7;
+  size_t r8;
+  size_t r9;
+  size_t r10;
+  size_t r11;
+  size_t r12;
   /* stack pointer is in the thread struct */
-  uint32_t lr;
+  size_t lr;
   /* aka the exception mode lr */
-  uint32_t pc;
-  uint32_t cpsr;
+  size_t pc;
+  size_t cpsr;
   /* patched in by the kernel */
-  uint32_t sp;
+  size_t sp;
 } __attribute__((packed)) RegisterContext;
 // To tell the kernel that everything but sp needs
 // to be written back to the thread's stack.
 // (since sp is part of the Thread struct itself)
-#define STACK_CTX_SIZE (sizeof(RegisterContext)-sizeof(uint32_t))
+#define STACK_CTX_SIZE (sizeof(RegisterContext)-sizeof(size_t))
 #endif
 
 void print_register_context(RegisterContext ctx);
